functions/authorize/requirements.txt
boto3==1.26.144
python-jose==3.3.0
